fix: some errors when building
