+++ client/src/app/faq/page.tsx
@@ -1,7 +1,5 @@
-"use client";
 import * as React from "react";
-import * as TabsPrimitive from "@radix-ui/react-tabs";
-import { cn } from "@/lib/utils";
+import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 
 const faqData = [
   {
@@ -50,74 +48,12 @@ const faqData = [
     texto1:
       "No canto inferior da tela principal, existe um footer com um índice de links rápidos. ",
     titulo2: "Passo 2",
-    texto2:
-      "Clique no link 'Sobre nós'.",
+    texto2: "Clique no link 'Sobre nós'.",
     titulo3: "Passo 3",
     texto3:
       "Nessa tela existe tudo explicado sobre nossa missão e nossa história, boa leitura!",
   },
 ];
-
-function Tabs({
-  className,
-  ...props
-}: React.ComponentProps<typeof TabsPrimitive.Root>) {
-  return (
-    <TabsPrimitive.Root
-      data-slot="tabs"
-      className={cn("flex flex-col gap-2", className)}
-      {...props}
-    />
-  );
-}
-
-function TabsList({
-  className,
-  ...props
-}: React.ComponentProps<typeof TabsPrimitive.List>) {
-  return (
-    <TabsPrimitive.List
-      data-slot="tabs-list"
-      className={cn(
-        "bg-white border-white inline-flex flex-row md:flex-col flex-wrap md:h-fit items-center justify-start rounded-lg p-[3px] gap-1 overflow-x-auto md:overflow-visible",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-function TabsTrigger({
-  className,
-  ...props
-}: React.ComponentProps<typeof TabsPrimitive.Trigger>) {
-  return (
-    <TabsPrimitive.Trigger
-      data-slot="tabs-trigger"
-      className={cn(
-        "bg-white text-black data-[state=active]:bg-gray-100 data-[state=active]:text-blue-900",
-        "data-[state=active]:shadow-sm border border-transparent rounded-md px-3 py-1 text-sm font-medium transition-colors w-[120px] md:w-[150px]",
-        className
-      )}
-      {...props}
-    />
-  );
-}
-
-function TabsContent({
-  className,
-  ...props
-}: React.ComponentProps<typeof TabsPrimitive.Content>) {
-  return (
-    <TabsPrimitive.Content
-      data-slot="tabs-content"
-      className={cn("flex-1 outline-none", className)}
-      {...props}
-    />
-  );
-}
-
-export { Tabs, TabsList, TabsTrigger, TabsContent };
 
 export default function Faq() {
   return (
