+++ client/src/app/home/create-activity/page.tsx
@@ -47,7 +47,6 @@ export default function CreateActionPage() {
 
   const [tags, setTags] = useState<string[]>([]);
   const [tagInput, setTagInput] = useState("");
-  const [previews, setPreviews] = useState<string[]>([]);
 
   function handleChange(
     e: React.ChangeEvent<
@@ -94,7 +93,6 @@ export default function CreateActionPage() {
         endDate: "",
       });
       setTags([]);
-      setPreviews([]);
     }
   }, [state?.ok, router]);
