+++ client/src/components/activity/Activity.tsx
@@ -23,8 +23,6 @@ export default function ActivityComp({
   user,
   initialActivity,
 }: ActivityProps) {
-  if (!initialActivity) return null;
-
   const [activity, setActivity] = React.useState<Activity | null>(
     initialActivity
   );
