+++ client/src/components/buttons/LargeButton.tsx
@@ -1,6 +1,5 @@
 import React from "react";
 
-// eslint-disable-next-line @typescript-eslint/no-empty-object-type
 interface LargeButtonProps
   extends React.ButtonHTMLAttributes<HTMLButtonElement> {
   variant?: "primary" | "secondary";
