+++ client/src/components/homes/OngHome.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import { useState } from "react";
-import LargeButton from "@/components/buttons/LargeButton";
 import ActivityCard from "@/components/general/ActivityCard";
 import SetLoading from "@/components/setLoading/setLoading";
 import Link from "next/link";
