+++ client/src/components/homes/SchoolHome.tsx
@@ -3,12 +3,11 @@ import { AuthUser } from "@/actions/getAuthUser";
 import ActivityCard from "@/components/general/ActivityCard";
 import SetLoading from "@/components/setLoading/setLoading";
 import React, { useState } from "react";
-import getSchoolActivitiesBySchoolId, {
-  SchoolActivity,
-} from "@/actions/getSchoolActivity";
+import getSchoolActivitiesBySchoolId from "@/actions/getSchoolActivity";
+import { Activity } from "@/actions/getActivities";
 
 export default function SchoolHome({ authUser }: { authUser: AuthUser }) {
-  const [activities, setActivities] = useState<SchoolActivity[] | null>([]);
+  const [activities, setActivities] = useState<Activity[] | null>([]);
 
   React.useEffect(() => {
     const fetchActivities = async () => {
