+++ client/src/components/settings/SettingsCard.tsx
@@ -1,4 +1,3 @@
-import Link from "next/link";
 import { IconType } from "react-icons";
 
 interface SettingsCardProps {
